+++ Realm/RLMObject.h
@@ -19,6 +19,7 @@
 #import <Foundation/Foundation.h>
 
 #import <Realm/RLMObjectBase.h>
+#import <Realm/RLMThreadSafeReference.h>
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -91,7 +92,7 @@ NS_ASSUME_NONNULL_BEGIN
     times when you *cannot* begin a write transaction.
  */
 
-@interface RLMObject : RLMObjectBase
+@interface RLMObject : RLMObjectBase <RLMThreadConfined>
 
 #pragma mark - Creating & Initializing Objects