+++ RealmSwift/RealmCollection.swift
@@ -124,7 +124,8 @@ public enum RealmCollectionChange<T> {
 /**
  A homogenous collection of `Object`s which can be retrieved, filtered, sorted, and operated upon.
 */
-public protocol RealmCollection: RandomAccessCollection, LazyCollectionProtocol, CustomStringConvertible {
+public protocol RealmCollection: RandomAccessCollection, LazyCollectionProtocol, CustomStringConvertible, ThreadConfined {
+    // Must also conform to `AssistedObjectiveCBridgeable`
 
     /// The type of the objects contained in the collection.
     associatedtype Element: Object
@@ -367,7 +368,7 @@ public protocol RealmCollection: RandomAccessCollection, LazyCollectionProtocol,
     func _addNotificationBlock(_ block: @escaping (RealmCollectionChange<AnyRealmCollection<Element>>) -> Void) -> NotificationToken
 }
 
-private class _AnyRealmCollectionBase<T: Object> {
+private class _AnyRealmCollectionBase<T: Object>: AssistedObjectiveCBridgeable {
     typealias Wrapper = AnyRealmCollection<Element>
     typealias Element = T
     var realm: Realm? { fatalError() }
@@ -396,6 +397,8 @@ private class _AnyRealmCollectionBase<T: Object> {
     func setValue(_ value: Any?, forKey key: String) { fatalError() }
     func _addNotificationBlock(_ block: @escaping (RealmCollectionChange<Wrapper>) -> Void)
         -> NotificationToken { fatalError() }
+    class func bridging(from objectiveCValue: Any, with metadata: Any?) -> Self { fatalError() }
+    var bridged: (objectiveCValue: Any, metadata: Any?) { fatalError() }
 }
 
 private final class _AnyRealmCollection<C: RealmCollection>: _AnyRealmCollectionBase<C.Element> {
@@ -500,6 +503,17 @@ private final class _AnyRealmCollection<C: RealmCollection>: _AnyRealmCollection
     /// :nodoc:
     override func _addNotificationBlock(_ block: @escaping (RealmCollectionChange<Wrapper>) -> Void)
         -> NotificationToken { return base._addNotificationBlock(block) }
+
+    // MARK: AssistedObjectiveCBridgeable
+
+    override class func bridging(from objectiveCValue: Any, with metadata: Any?) -> _AnyRealmCollection {
+        return _AnyRealmCollection(
+            base: (C.self as! AssistedObjectiveCBridgeable.Type).bridging(from: objectiveCValue, with: metadata) as! C)
+    }
+
+    override var bridged: (objectiveCValue: Any, metadata: Any?) {
+        return (base as! AssistedObjectiveCBridgeable).bridged
+    }
 }
 
 /**
@@ -514,7 +528,11 @@ public final class AnyRealmCollection<T: Object>: RealmCollection {
 
     /// The type of the objects contained in the collection.
     public typealias Element = T
-    private let base: _AnyRealmCollectionBase<T>
+    fileprivate let base: _AnyRealmCollectionBase<T>
+
+    fileprivate init(base: _AnyRealmCollectionBase<T>) {
+        self.base = base
+    }
 
     /// Creates an `AnyRealmCollection` wrapping `base`.
     public init<C: RealmCollection>(_ base: C) where C.Element == T {
@@ -803,6 +821,26 @@ public final class AnyRealmCollection<T: Object>: RealmCollection {
         -> NotificationToken { return base._addNotificationBlock(block) }
 }
 
+// MARK: AssistedObjectiveCBridgeable
+
+private struct AnyRealmCollectionBridgingMetadata<T: Object> {
+    var baseMetadata: Any?
+    var baseType: _AnyRealmCollectionBase<T>.Type
+}
+
+extension AnyRealmCollection: AssistedObjectiveCBridgeable {
+    static func bridging(from objectiveCValue: Any, with metadata: Any?) -> AnyRealmCollection {
+        guard let metadata = metadata as? AnyRealmCollectionBridgingMetadata<T> else { preconditionFailure() }
+        return AnyRealmCollection(base: metadata.baseType.bridging(from: objectiveCValue, with: metadata.baseMetadata))
+    }
+
+    var bridged: (objectiveCValue: Any, metadata: Any?) {
+        return (
+            objectiveCValue: base.bridged.objectiveCValue,
+            metadata: AnyRealmCollectionBridgingMetadata(baseMetadata: base.bridged.metadata, baseType: type(of: base))
+        )
+    }
+}
 
 // MARK: Unavailable
 
@@ -939,7 +977,8 @@ public enum RealmCollectionChange<T> {
  A homogenous collection of `Object`s which can be retrieved, filtered, sorted,
  and operated upon.
 */
-public protocol RealmCollectionType: CollectionType, CustomStringConvertible {
+public protocol RealmCollectionType: CollectionType, CustomStringConvertible, ThreadConfined {
+    // Must also conform to `AssistedObjectiveCBridgeable`
 
     /// The type of the objects contained in the collection.
     associatedtype Element: Object
@@ -1171,7 +1210,7 @@ public protocol RealmCollectionType: CollectionType, CustomStringConvertible {
     func _addNotificationBlock(block: (RealmCollectionChange<AnyRealmCollection<Element>>) -> Void) -> NotificationToken
 }
 
-private class _AnyRealmCollectionBase<T: Object> {
+private class _AnyRealmCollectionBase<T: Object>: AssistedObjectiveCBridgeable {
     typealias Wrapper = AnyRealmCollection<Element>
     typealias Element = T
     var realm: Realm? { fatalError() }
@@ -1200,6 +1239,8 @@ private class _AnyRealmCollectionBase<T: Object> {
     func setValue(value: AnyObject?, forKey key: String) { fatalError() }
     func _addNotificationBlock(block: (RealmCollectionChange<Wrapper>) -> Void)
         -> NotificationToken { fatalError() }
+    class func bridging(from objectiveCValue: Any, with metadata: Any?) -> Self { fatalError() }
+    var bridged: (objectiveCValue: Any, metadata: Any?) { fatalError() }
 }
 
 private final class _AnyRealmCollection<C: RealmCollectionType>: _AnyRealmCollectionBase<C.Element> {
@@ -1294,6 +1335,17 @@ private final class _AnyRealmCollection<C: RealmCollectionType>: _AnyRealmCollec
     /// :nodoc:
     override func _addNotificationBlock(block: (RealmCollectionChange<Wrapper>) -> Void)
         -> NotificationToken { return base._addNotificationBlock(block) }
+
+    // MARK: AssistedObjectiveCBridgeable
+
+    override class func bridging(from objectiveCValue: Any, with metadata: Any?) -> _AnyRealmCollection {
+        return _AnyRealmCollection(
+            base: (C.self as! AssistedObjectiveCBridgeable.Type).bridging(from: objectiveCValue, with: metadata) as! C)
+    }
+
+    override var bridged: (objectiveCValue: Any, metadata: Any?) {
+        return (base as! AssistedObjectiveCBridgeable).bridged
+    }
 }
 
 /**
@@ -1307,6 +1359,10 @@ public final class AnyRealmCollection<T: Object>: RealmCollectionType {
     /// The type of the objects contained in the collection.
     public typealias Element = T
     private let base: _AnyRealmCollectionBase<T>
+
+    private init(base: _AnyRealmCollectionBase<T>) {
+        self.base = base
+    }
 
     /// Creates an `AnyRealmCollection` wrapping `base`.
     public init<C: RealmCollectionType where C.Element == T>(_ base: C) {
@@ -1572,6 +1628,27 @@ public final class AnyRealmCollection<T: Object>: RealmCollectionType {
     /// :nodoc:
     public func _addNotificationBlock(block: (RealmCollectionChange<AnyRealmCollection>) -> Void)
         -> NotificationToken { return base._addNotificationBlock(block) }
+}
+
+// MARK: AssistedObjectiveCBridgeable
+
+private struct AnyRealmCollectionBridgingMetadata<T: Object> {
+    var baseMetadata: Any?
+    var baseType: _AnyRealmCollectionBase<T>.Type
+}
+
+extension AnyRealmCollection: AssistedObjectiveCBridgeable {
+    static func bridging(from objectiveCValue: Any, with metadata: Any?) -> AnyRealmCollection {
+        guard let metadata = metadata as? AnyRealmCollectionBridgingMetadata<T> else { preconditionFailure() }
+        return AnyRealmCollection(base: metadata.baseType.bridging(from: objectiveCValue, with: metadata.baseMetadata))
+    }
+
+    var bridged: (objectiveCValue: Any, metadata: Any?) {
+        return (
+            objectiveCValue: base.bridged.objectiveCValue,
+            metadata: AnyRealmCollectionBridgingMetadata(baseMetadata: base.bridged.metadata, baseType: base.dynamicType)
+        )
+    }
 }
 
 #endif